requirements.txt
Fabric3==1.14.post1
Jinja2==2.10
peewee==3.2.3
